webpack for server bundle

+++ src/server/index.js
@@ -1,25 +1,41 @@
 const {register, verifyRegistration, authenticate} = require('./identityBroker');
 const {putToDo, getToDo, deleteToDo, fetchToDos} = require('./dataAccessorBroker');
 
-export const handler = (event) => {
+// eslint-disable-next-line complexity
+const handler = async (event, context, callback) => {
     const operation = event.operation;
 
-    switch (operation) {
-        case 'REGISTER':
-            return register(event.body);
-        case 'VERIFY_REGISTRATION':
-            return verifyRegistration(event.body);
-        case 'AUTHENTICATE':
-            return authenticate(event.body);
-        case 'PUT_TODO':
-            return putToDo(event.body);
-        case 'GET_TODO':
-            return getToDo(event.body);
-        case 'DELETE_TODO':
-            return deleteToDo(event.body);
-        case 'FETCH_TODOS':
-            return fetchToDos(event.body);
-        default:
-            return `[BAD REQUEST] Unknown operation: ${operation}`;
+    if (operation === 'REGISTER') {
+        const response = await register(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'VERIFY_REGISTRATION') {
+        const response = await verifyRegistration(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'AUTHENTICATE') {
+        const response = await authenticate(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'PUT_TODO') {
+        const response = await putToDo(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'GET_TODO') {
+        const response = await getToDo(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'DELETE_TODO') {
+        const response = await deleteToDo(event.body);
+
+        return callback(null, response);
+    } else if (operation === 'FETCH_TODOS') {
+        const response = await fetchToDos(event.body);
+
+        return callback(null, response);
     }
+
+    return callback(new Error(`[BAD REQUEST] Unknown operation: ${operation}`));
 };
+
+module.exports = {handler};
